Use ^P / ^N for beginning-/substring-search bindings
The previous config used ^O (vi-like, seems fine) and ^I (never worked,
<Tab> is expand-or-complete).

^P and ^N were on up-/down-line-or-history, which are still available
with default bindings in viins on <Up> and <Down> and in vicmd on k and j.
And besides, history-beginning-search-backward/-forward go up and down
when the cursor is at the beginning of the line, so this doesn't take
too much adjustment to hit ^A first if needed.