Don't run reset-prompt during zle-line-init
This partially addresses issue #5, keeping the %_ data intact so long as
ZLE stays in viins mode. But there's just no way to preserve it and
update the prompt. There are two possible further changes to try:

1. Fix ZSH so %_ always references the top level cmdstack instead
instead of finding the one that ZLE installs. This obviously is the best
choice and the whole problem would go away.

2. Have an option to disable calling restore-prompt when R/PS2 is being
shown (where %_ and %^ are useful). I don't see any obvious indicators
to test for that inside a ZLE hook, but it's possible there's some way
to get at it. Then users could pick to either have the mode indicator
displayed OR the %_ text.