Hew more closely to default search & up-line behavior
In reviewing changes made to various forks of this project, it appears
that history-beginning-search-forward and -backward are not pleasant to
use for Down and Up actions. Similarly, the incremental search without
patterns is what Emacs mode uses and what will likely be preferred.

Anyone who wants the previous behavior can add the following to .zshrc
after loading this module:

    vim-mode-bindkey viins vicmd -- history-incremental-pattern-search-backward '^R'
    vim-mode-bindkey viins vicmd -- history-incremental-pattern-search-forward  '^S'

    vim-mode-bindkey viins vicmd -- history-beginning-search-backward   '^P'
    vim-mode-bindkey viins vicmd -- history-beginning-search-forward    '^N'
    vim-mode-bindkey viins vicmd -- history-beginning-search-backward   Up
    vim-mode-bindkey viins vicmd -- history-beginning-search-forward    Down