Work around Application-mode problem with Home/End keys
On some terminals (Alacritty, iTerm2, kitty, probably others) the
bindings the terminal uses depends on if it's in Application mode or
not. See the comment at the top of the Special Keys section for why
this hasn't been dealt with earlier. Clearly there were problems that
simply didn't show up with my usage.

Fixes #16. See #22 for the underlying problem.