Ensure zle hook functions return true
The documentation is a bit sparse for this, but says:

    The return status reflects the success or failure of the operation
    carried out by the widget, or if it is a user-defined widget the
    return status of the shell function.

    A non-zero return status causes the shell to beep when the widget
    exits, unless the BEEP options was unset or the widget was called
    via the zle command.  Thus if a user defined widget requires an
    immediate beep, it should call the beep widget directly.

Anyways, it was `vim-mode-handle-event()`, which was returning false from
its `line-init` branch, which triggered this blank cursor.

I suspect there's a bug in powerlevel10k too, which makes it sensitive
to some arbitrary other hook's return value. Regardless, it's probably
good to ensure every widget returns true if it's happy.

Closes #38